requirements.txt
llama-cpp-python==0.2.63
pyttsx3==2.90
speechrecognition==3.9.0
